Fix download manager displayed odd size when downloading huge file.
Detail:
  1.QProgressBar's properties(maximum,value) don't support large value(over int_max),
    thus, this fix changed to percentage from file recept size and file size.

  2.Add support Giga bytes.

  See also issue: 418.

+++ src/downloadmanager.h
@@ -149,7 +149,7 @@ public:
     void setRemovePolicy(RemovePolicy policy);
 
     static QString timeString(double timeRemaining);
-    static QString dataString(int size);
+    static QString dataString(qint64 size);
 
 public slots:
     void download(const QNetworkRequest &request, bool requestFileName = false);
